Fix #62
Move hanging signs to OptiGlue
